start of refactor to make faster and more reliable
