--- events/join-vatsim.js
@@ -1,21 +1,0 @@
-const { ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js')
-
-module.exports = {
-  name: 'messageCreate',
-  once: false,
-  execute (message) {
-    if (message.content.includes('join vatsim')) {
-      const row = new ActionRowBuilder()
-        .addComponents(
-          new ButtonBuilder()
-            .setLabel('Joining VATSIM')
-            .setURL('https://www.thepilotclub.org/resources#VATSIM')
-            .setStyle(ButtonStyle.Link)
-        )
-      message.reply({
-        content: 'To Join VATSIM you should go to this website and click register!',
-        components: [row]
-      })
-    }
-  }
-}
