+++ package.json
@@ -2,6 +2,7 @@
   "name": "discord-bot",
   "version": "1.0.0",
   "description": "",
+  "type": "module",
   "scripts": {
     "start": "node bot.js",
     "start:dev": "nodemon bot.js",
